adapt queries from PlannerSpec
Also move some queries to PlannerSql2ExactSpec
