+++ mongodb/src/test/resources/planner/plan_double_aggregation_with_another_projection.txt
@@ -37,7 +37,7 @@ Chain
 │  │  │  │          }
 │  │  │  │        })
 │  │  │  ╰─ By
-│  │  │     ╰─ Name("0" -> ["$foo"])
+│  │  │     ╰─ Name("0" -> ["$state"])
 │  │  ├─ $SimpleMapF
 │  │  │  ├─ Map
 │  │  │  │  ╰─ JsCore([[_._id["0"]], _.f0, [[_._id["0"]], _.f1]])
@@ -94,7 +94,7 @@ Chain
 │     │  │          }
 │     │  │        })
 │     │  ╰─ By
-│     │     ╰─ Name("0" -> ["$foo"])
+│     │     ╰─ Name("0" -> ["$state"])
 │     ├─ $SimpleMapF
 │     │  ├─ Map
 │     │  │  ╰─ JsCore([[_._id["0"]], _.f0, [[_._id["0"]], _.f1]])
