+++ mongodb/src/test/resources/planner/plan_simple_join_(map-reduce).txt
@@ -32,7 +32,7 @@ Chain
 │  │     │  ╰─ Name("_id" -> "$_id")
 │  │     ╰─ IncludeId
 │  ╰─ Chain
-│     ├─ $ReadF(db; zips2)
+│     ├─ $ReadF(db; smallZips)
 │     ├─ $SimpleMapF
 │     │  ├─ Map
 │     │  │  ╰─ Let(__val)
